# [core] Optimize dynamic storage allocator