Closes #407.

## Benches for 100k `BitStash::put` ops (updated)

### New version (no more loop):

```
Bench: populated cache/fill_bitstash                                                                             
                        time:   [16.253 ms 16.922 ms 17.659 ms]                                                                                                
                        change: [-10.013% -3.6306% +3.2150%] (p = 0.29 > 0.05)                                                                                 
                        No change in performance detected.                     
Found 13 outliers among 100 measurements (13.00%)                              
  5 (5.00%) high mild                                                                                                                                          
  8 (8.00%) high severe                                                                                                                                        
Bench: populated cache/one_put                                                                             
                        time:   [465.78 ns 500.94 ns 538.52 ns]                
                        change: [-97.830% -97.481% -97.089%] (p = 0.00 < 0.05)                                                                                 
                        Performance has improved.                                                                                                              
Found 3 outliers among 100 measurements (3.00%)                                                                                                                
  3 (3.00%) high mild                                                                                                                                          
                                                                                                                                                               
Bench: empty cache/fill_bitstash                                                                                                                               
                        time:   [17.082 ms 17.946 ms 18.910 ms]                                                                                                
                        change: [-1.5820% +5.6911% +13.130%] (p = 0.11 > 0.05)                                                                                 
                        No change in performance detected.
Found 12 outliers among 100 measurements (12.00%)
  7 (7.00%) high mild
  5 (5.00%) high severe
Bench: empty cache/one_put                                                                             
                        time:   [7.7743 us 8.3617 us 9.0137 us]
                        change: [-5.6624% +3.8722% +14.787%] (p = 0.47 > 0.05)
                        No change in performance detected.
Found 18 outliers among 100 measurements (18.00%)
  5 (5.00%) high mild
  13 (13.00%) high severe
```

### Running the old version (loop) subsequently after the new version:
```
Bench: populated cache/fill_bitstash                                                                            
                        time:   [28.347 ms 29.501 ms 30.765 ms]
                        change: [+64.246% +74.338% +85.703%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 13 outliers among 100 measurements (13.00%)
  2 (2.00%) high mild
  11 (11.00%) high severe
Bench: populated cache/one_put                                                                            
                        time:   [757.24 ns 810.25 ns 867.63 ns]
                        change: [+46.578% +61.744% +77.221%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

Bench: empty cache/fill_bitstash                                                                            
                        time:   [27.392 ms 28.532 ms 29.816 ms]
                        change: [+48.523% +58.986% +69.641%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 10 outliers among 100 measurements (10.00%)
  2 (2.00%) high mild
  8 (8.00%) high severe
Bench: empty cache/one_put                                                                            
                        time:   [8.2094 us 8.9335 us 9.7604 us]
                        change: [-4.1931% +6.8383% +21.234%] (p = 0.27 > 0.05)
                        No change in performance detected.
Found 19 outliers among 100 measurements (19.00%)
  1 (1.00%) high mild
  18 (18.00%) high severe
```